amazing changes, added deploy module, add and delete.

+++ fuel/app/classes/controller/api/ftp.php
@@ -6,41 +6,84 @@ class Controller_Api_Ftp extends Controller {
         
     }
 
-    public function action_getall() {
-
+    /**
+     * List one or more FTP.
+     * @param type $id
+     * @return type
+     */
+    public function action_getall($id = null) {
         if (!Auth::check()) {
             echo json_encode(array(
                 'status' => false,
-                'reason' => 'Not logged in'
+                'reason' => 'Not logged in',
+                'request' => Input::post()
             ));
             return;
         }
 
         $user_id = Auth::get_user_id()[1];
-        $data = DB::select()->from('ftpdata')->where('user_id', $user_id)->execute()->as_array();
+        $data = DB::select()->from('ftpdata')->where('user_id', $user_id);
+
+        if ($id !== null) {
+            $data = $data->and_where('id', $id);
+        }
+
+        $data = $data->execute()->as_array();
+
         echo json_encode(array(
             'status' => true,
             'data' => $data
         ));
     }
 
+    /**
+     * adding a FTP server.
+     * @return boolean
+     */
     public function action_addftp() {
         if (!Auth::check()) {
-            return false;
+
+            echo json_encode(array(
+                'status' => false,
+                'reason' => 'Not logged in',
+                'request' => Input::post()
+            ));
+            return;
+            
         }
 
         $data = Input::post();
         $user_id = Auth::get_user_id()[1];
         $data['user_id'] = $user_id;
-        $a = DB::insert('ftpdata')->set($data)->execute();
-        if ($a) {
+
+        $existing = DB::select()->from('ftpdata')
+                        ->where('host', $data['host'])
+                        ->and_where('username', $data['username'])
+                        ->and_where('user_id', $user_id)
+                        ->execute()->as_array();
+
+        if (count($existing) > 0) {
             echo json_encode(array(
-                'status' => true,
-                'request' => Input::post()
+                'status' => false,
+                'request' => Input::post(),
+                'reason' => 'A FTP account with the same host and username already exist.'
             ));
+        } else {
+            $a = DB::insert('ftpdata')->set($data)->execute();
+            if ($a) {
+                echo json_encode(array(
+                    'status' => true,
+                    'request' => Input::post()
+                ));
+            }
         }
     }
 
+    /**
+     * Delete a FTP server by ID
+     * @param type $id
+     * @return type
+     */
     public function action_delftp($id) {
 
         if (!Auth::check()) {
@@ -52,18 +95,33 @@ class Controller_Api_Ftp extends Controller {
             return;
         }
 
-        $result = DB::delete('ftpdata')->where('id', $id)->execute();
+        $row = DB::select()->from('ftpdata')
+                        ->where('id', $id)
+                        ->and_where('user_id', Auth::get_user_id()[1])
+                        ->execute()->as_array();
 
-        if ($result) {
-            echo json_encode(array(
-                'status' => true,
-                'request' => Input::post()
-            ));
-        } else {
+        if (count($row) == 0) {
             echo json_encode(array(
                 'status' => false,
-                'request' => Input::post()
+                'request' => Input::post(),
+                'reason' => 'You are not authorized to delete.'
             ));
+        } else {
+
+            $result = DB::delete('ftpdata')->where('id', $id)->execute();
+
+            if ($result) {
+                echo json_encode(array(
+                    'status' => true,
+                    'request' => Input::post()
+                ));
+            } else {
+                echo json_encode(array(
+                    'status' => false,
+                    'request' => Input::post(),
+                    'reason' => 'Cound not insert the value'
+                ));
+            }
         }
     }
